S15.f.10: extract plugin.ts OpenClaw command/provider/CLI registrations (3233 → 2463 LOC, §4.2 cap reached)
Final S15.f sub-slice. Extracts the Foundry api.registerProvider call,
the api.registerCli registrar emitting 'openclaw azureclaw {status,
connect,dev,logs}', and the ~12 api.registerCommand slash-command
('/azureclaw …') definitions to cli/src/core/commands/openclaw.ts.

Tool/command bodies are byte-identical; only closure capture is
replaced with explicit OpenClawCommandsDeps threading (log, config,
getFoundryProject, meshClient, identity, policy, trustStore,
auditLogger, memorySyncBuffer, syncToFoundryMemory).

cli/src/plugin.ts: 3233 → 2463 LOC (−770).
Cumulative S15.f: 7139 → 2463 LOC (−4676).
§4.2 Phase 2 cap of 3000 LOC achieved (537 LOC under cap).

tsc clean / lint 33 warnings / 454 tests pass / build clean.
